easings numberseting parameter

+++ src/main/java/club/lyric/infinity/api/setting/settings/IntegerSetting.java
@@ -27,7 +27,7 @@ public class IntegerSetting extends Setting<Integer> {
     @Override
     public Integer getMax()
     {
-        return (int)max;
+        return max;
     }
 
     @Override
